requirements.txt
beautifulsoup4==4.9.3
certifi==2022.5.18.1
chardet==4.0.0
googlesearch-python==1.1.0
idna==2.10
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.1
requests==2.25.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
